+++ tests/test_reporting_utils.py
@@ -4,7 +4,7 @@ import pytest
 import xarray as xr
 
 from ixmp.reporting import Key
-from ixmp.reporting.utils import AttrSeries, Quantity
+from ixmp.reporting.utils import AttrSeries, Quantity, as_quantity
 from ixmp.testing import assert_qty_allclose, assert_qty_equal
 
 
@@ -94,13 +94,55 @@ class TestQuantity:
 
 class TestAttrSeries:
     """Tests of AttrSeries in particular."""
-    def test_sum(self):
+    @pytest.fixture
+    def foo(self):
         idx = pd.MultiIndex.from_product([['a1', 'a2'], ['b1', 'b2']],
                                          names=['a', 'b'])
-        foo = AttrSeries([0, 1, 2, 3], index=idx)
+        yield AttrSeries([0, 1, 2, 3], index=idx)
 
+    def test_sum(self, foo):
         # AttrSeries can be summed across all dimensions
         result = foo.sum(dim=['a', 'b'])
         assert isinstance(result, AttrSeries)  # returns an AttrSeries
         assert len(result) == 1                # with one element
         assert result[0] == 6                  # that has the correct value
+
+    def test_others(self, foo):
+        # Exercise other compatibility functions
+        assert isinstance(foo.as_xarray(), xr.DataArray)
+        assert type(foo.to_frame()) is pd.DataFrame
+        assert foo.drop('a').dims == ('b',)
+
+
+@pytest.mark.skipif(Quantity is AttrSeries, reason='For xr.DataArray only')
+def test_as_quantity():
+    """Test conversion to sparse.COO-backed xr.DataArray in as_quantity()."""
+    x_series = pd.Series(
+        data=[1, 2, 3, 4],
+        index=pd.MultiIndex.from_product([['a', 'b'], ['c', 'd']],
+                                         names=['foo', 'bar']),
+    )
+    y_series = pd.Series(data=[5, 6], index=pd.Index(['e', 'f'], name='baz'))
+
+    x = xr.DataArray.from_series(x_series, sparse=True)
+    y = xr.DataArray.from_series(y_series, sparse=True)
+
+    x_dense = xr.DataArray.from_series(x_series)
+    y_dense = xr.DataArray.from_series(y_series)
+
+    with pytest.raises(ValueError, match='make sure that the broadcast shape'):
+        x_dense * y
+
+    z1 = as_quantity(x_dense) * y
+    z2 = x * as_quantity(y_dense)
+    assert z1.dims == ('foo', 'bar', 'baz')
+    assert_qty_equal(z1, z2)
+
+    z3 = as_quantity(x) * as_quantity(y)
+    assert_qty_equal(z1, z3)
+
+    z4 = as_quantity(x) * y
+    assert_qty_equal(z1, z4)
+
+    z5 = as_quantity(x_series) * y
+    assert_qty_equal(z1, z5)